+++ code/botlib/be_ai_weight.c
@@ -345,7 +345,7 @@ weightconfig_t *ReadWeightConfig(char *filename)
 				return NULL;
 			} //end if
 			StripDoubleQuotes(token.string);
-			config->weights[config->numweights].name = (char *) GetClearedMemory(strlen(token.string) + 1);
+			config->weights[config->numweights].name = (char *) GetClearedMemory((int)strlen(token.string) + 1);
 			strcpy(config->weights[config->numweights].name, token.string);
 			if (!PC_ExpectAnyToken(source, &token))
 			{
